improve: using generic methods to generate healthEndpointPath and swaggerEndpointPath
