+++ pkg/utils/utils.go
@@ -1,3 +1,4 @@
+// Package utils provides some common utils.
 package utils
 
 import (
